Move copy on write list to spearate file
